+++ src/ExportManager/PrintSettings.cs
@@ -274,7 +274,7 @@ namespace SCaddins.ExportManager
                 pm.PrintSetup.InSession.PrintParameters.PaperPlacement = PaperPlacementType.Margins;
                 pm.PrintSetup.CurrentPrintSetting.PrintParameters.MarginType = MarginType.UserDefined;
                 pm.PrintSetup.InSession.PrintParameters.MarginType = MarginType.UserDefined;
-#if REVIT2022 || REVIT2023
+#if REVIT2022 || REVIT2023 || REVIT2024
                 pm.PrintSetup.CurrentPrintSetting.PrintParameters.OriginOffsetX = 0;
                 pm.PrintSetup.InSession.PrintParameters.OriginOffsetY = 0;
 #else