+++ src/ExportManager/ViewModels/OptionsViewModel.cs
@@ -339,7 +339,7 @@ namespace SCaddins.ExportManager.ViewModels
 
         public bool ExportAdobePDFEnabled
         {
-#if REVIT2022 || REVIT2023
+#if REVIT2022 || REVIT2023 || REVIT2024
             get { return false; }
 #else
             get { return true; }
@@ -403,7 +403,7 @@ namespace SCaddins.ExportManager.ViewModels
 
         public bool ExportRevitPDFEnabled
         {
-#if REVIT2022 || REVIT2023
+#if REVIT2022 || REVIT2023 || REVIT2024
             get { return true; }
 #else
             get { return false; }