+++ src/RevisionUtilities/RevisionUtilities.cs
@@ -127,7 +127,11 @@ namespace SCaddins.RevisionUtilities
                     data[cloudNumber, 4] = revCloud.Comments;
                     data[cloudNumber, 5] = revCloud.Date;
                     data[cloudNumber, 6] = revCloud.Description;
+#if REVIT2024
+                    data[cloudNumber, 7] = revCloud.Id.Value.ToString(CultureInfo.InvariantCulture);
+#else
                     data[cloudNumber, 7] = revCloud.Id.IntegerValue.ToString(CultureInfo.InvariantCulture);
+#endif
                 }
             }