+++ src/SolarAnalysis/SolarAnalysisManager.cs
@@ -353,7 +353,7 @@ namespace SCaddins.SolarAnalysis
                 {
                     var view = (View)element;
                     var v = view;
-#if REVIT2019 || REVIT2020 || REVIT2021 || REVIT2022 || REVIT2023
+#if REVIT2019 || REVIT2020 || REVIT2021 || REVIT2022 || REVIT2023 || REVIT2024
                     if (v.Name == name)
                     {
                         return false;